+++ lib/models/user_data.dart
@@ -14,6 +14,11 @@ class UserData {
   Color appColor;
   Map<String, Map<String, List<Map<String, dynamic>>>> foodDataByDate;
   List<String> fcmTokens;
+  int? caloriesGoal;
+  int? proteinGoal;
+  int? carbsGoal;
+  int? fatGoal;
+  String? weightGoalType;
 
   // constructor
   UserData({
@@ -27,8 +32,15 @@ class UserData {
     this.appColor = const Color.fromARGB(255, 45, 45, 45), // default app color
     Map<String, Map<String, List<Map<String, dynamic>>>>? foodDataByDate,
     String? username,
-    this.reminders = const [],
-    this.fcmTokens = const [],
+    List<ReminderData>? reminders,
+    List<String>? fcmTokens,
+    this.caloriesGoal,
+    this.proteinGoal,
+    this.carbsGoal,
+    this.fatGoal,
+    this.weightGoalType,
   }) : foodDataByDate = foodDataByDate ?? {},
+       reminders = reminders ?? [],
+       fcmTokens = fcmTokens ?? [],
        username = username ?? uid; // Default username is the UID
 }